requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.7
djangorestframework==3.8.2
idna==2.7
psycopg2==2.7.5
pydash==4.7.3
pytz==2018.5
requests==2.19.1
urllib3==1.23
